+++ cms/plugin_rendering.py
@@ -429,7 +429,7 @@ class ContentRenderer(BaseRenderer):
                 placeholder,
                 context=context,
                 language=language,
-                page=placeholder.source,
+                page=current_page,
                 editable=editable,
                 use_cache=True,
                 nodelist=None,